compiler: move dataviewcompiler to class
